feat: implement shipping label functionality with data retrieval
